Fix gravity (#15)
Gravity works again. All entities that interact gravitationally should be given a gravitational_mass value.
The baseline for this is that the player has a gravity of 1.
I opened all of the files and fixed most of the obvious global variables as well.
Also improved the debug mode in systems/DrawEntities.lua.